Add reactive email verification notice on editor
Add verification polling and signaling functions to userUtil

Use global user and userUtils instead of a prop

Fix broken email verification notice code

Move verification polling to user

Add verification polling and signaling functions to userUtil

Use global user and userUtils instead of a prop

Fix broken email verification notice code

Move verification polling to user

Refactor email-verification-notice

Add verified notice.

Add email verified notice to layout

Update i18n require.

Use a global notice for verification

Move verification check to user utils.

Add email verification notice to EditorGroundControl.

Add email verification polling.

Set a local storage value to signal email verification.

Show dialog upon clicking on notice.

Fix dialog on mobile devices.

Make notice match the button action.

Remove try again button.

Remove post saved message, add verification reasoning.

Disable the publish button if unverified.

Add a user mock to fix tests, add a new test case.

Tweaking the confirmation dialog.

Center notice tip.

Do not show focus ring on click.

Make sure email is clipped if too long.

Add resend email button.

Tweak copy.

Change button label to OK.

Fix resend email button.

Changing the word “verify” to “confirm” in the editor notice.

Add spinner when sending email.

Remove unused dependency.

Use user's new reactive email verification logic

Use new i18n import

Remove onPublish checks

Remove verification-notice styles entirely

Remove unused references

Remove i18n stub code

Use new wording for notice

Hide publish notice behind feature flag
