+++ client/post-editor/editor-ground-control/index.jsx
@@ -3,7 +3,8 @@
  */
 const noop = require( 'lodash/noop' ),
 	React = require( 'react' ),
-	PureRenderMixin = require( 'react-pure-render/mixin' );
+	PureRenderMixin = require( 'react-pure-render/mixin' ),
+	i18n = require( 'i18n-calypso' );
 
 /**
  * Internal dependencies
@@ -16,11 +17,14 @@ const Card = require( 'components/card' ),
 	StatusLabel = require( 'post-editor/editor-status-label' ),
 	postUtils = require( 'lib/posts/utils' ),
 	siteUtils = require( 'lib/site/utils' ),
+	userUtils = require( 'lib/user/utils' ),
 	PostSchedule = require( 'components/post-schedule' ),
 	postActions = require( 'lib/posts/actions' ),
 	Tooltip = require( 'components/tooltip' ),
 	PostListFetcher = require( 'components/post-list-fetcher' ),
-	stats = require( 'lib/posts/stats' );
+	stats = require( 'lib/posts/stats' ),
+	user = require( 'lib/user' )(),
+	config = require( 'config' );
 
 export default React.createClass( {
 	displayName: 'EditorGroundControl',
@@ -35,6 +39,7 @@ export default React.createClass( {
 		onPreview: React.PropTypes.func,
 		onPublish: React.PropTypes.func,
 		onSaveDraft: React.PropTypes.func,
+		onMoreInfoAboutEmailVerify: React.PropTypes.func,
 		post: React.PropTypes.object,
 		savedPost: React.PropTypes.object,
 		site: React.PropTypes.object,
@@ -59,13 +64,34 @@ export default React.createClass( {
 		};
 	},
 
+	componentDidMount: function() {
+		if ( ! config( 'email_verification_gate' ) ) return;
+		
+		user.on( 'change', this.updateNeedsVerification );
+		user.on( 'verify', this.updateNeedsVerification );
+	},
+
+	componentWillUnmount: function() {
+		if ( ! config( 'email_verification_gate' ) ) return;
+		
+		user.off( 'change', this.updateNeedsVerification );
+		user.off( 'verify', this.updateNeedsVerification );
+	},
+
+	updateNeedsVerification: function() {
+		this.setState( {
+			needsVerification: config( 'email_verification_gate' ) && userUtils.needsVerificationForSite( this.props.site ),
+		} );
+	},
+
 	getInitialState: function() {
 		return {
 			showSchedulePopover: false,
 			showAdvanceStatus: false,
 			showDateTooltip: false,
 			firstDayOfTheMonth: this.getFirstDayOfTheMonth(),
-			lastDayOfTheMonth: this.getLastDayOfTheMonth()
+			lastDayOfTheMonth: this.getLastDayOfTheMonth(),
+			needsVerification: config( 'email_verification_gate' ) && userUtils.needsVerificationForSite( this.props.site ),
 		};
 	},
 
@@ -145,6 +171,29 @@ export default React.createClass( {
 			publish: this.translate( 'Publish' ),
 			requestReview: this.translate( 'Submit for Review' ),
 		};
+		return buttonLabels[ primaryButtonState ];
+	},
+
+	getVerificationNoticeLabel: function() {
+		const primaryButtonState = this.getPrimaryButtonState();
+		let buttonLabels;
+		
+		// TODO: switch entirely to new wording once translations catch up
+		if ( i18n.getLocaleSlug() === 'en' ) {
+			buttonLabels = {
+				update: this.translate( 'To update, check your email and confirm your address.' ),
+				schedule: this.translate( 'To schedule, check your email and confirm your address.' ),
+				publish: this.translate( 'To publish, check your email and confirm your address.' ),
+				requestReview: this.translate( 'To submit for review, check your email and confirm your address.' ),
+			};
+		} else {
+			buttonLabels = {
+				update: this.translate( 'To update, please confirm your email address.' ),
+				schedule: this.translate( 'To schedule, please confirm your email address.' ),
+				publish: this.translate( 'To publish, please confirm your email address.' ),
+				requestReview: this.translate( 'To submit for review, please confirm your email address.' ),
+			};
+		}
 		return buttonLabels[ primaryButtonState ];
 	},
 
@@ -249,7 +298,8 @@ export default React.createClass( {
 	isPrimaryButtonEnabled: function() {
 		return ! this.props.isPublishing &&
 			! this.props.isSaveBlocked &&
-			this.props.hasContent;
+			this.props.hasContent &&
+			! this.state.needsVerification;
 	},
 
 	toggleAdvancedStatus: function() {
@@ -399,6 +449,17 @@ export default React.createClass( {
 						this.schedulePostPopover()
 					}
 				</div>
+				{
+					this.state.needsVerification &&
+					<div className="editor-ground-control__email-verification-notice" tabIndex={ 7 } onClick={ this.props.onMoreInfoAboutEmailVerify }>
+						<Gridicon
+							icon="info"
+							className="editor-ground-control__email-verification-notice-icon" />
+						{ this.getVerificationNoticeLabel() }
+						{ ' ' }
+						<u>{ this.translate( 'Learn More' ) }</u>
+					</div>
+				}
 			</Card>
 		);
 	}
