+++ client/post-editor/style.scss
@@ -1,9 +1,46 @@
-.dialog.verification-modal {
-	max-width: 475px;
+.dialog.confirmation-dialog {
+	max-width: 450px;
+	border-radius: 3px;
 
 	h1 {
+		font-size: 22px;
+		line-height: 1.2;
 		height: auto;
-		min-height: 2em;
+		margin: 0 0 4px 0;
+	}
+
+	p {
+		color: darken( $gray, 30 );
+		margin: 0 0 8px 0;
+
+		&.confirmation-dialog__email {
+			margin-bottom: 16px;
+			color: darken( $gray, 10 );
+
+			span {
+				display: inline-block;
+				max-width: 70%;
+				overflow: hidden;
+				text-overflow: ellipsis;
+				margin-right: 5px;
+			}
+		}
+
+		&.confirmation-dialog__reasoning {
+			color: darken( $gray, 10 );
+			font-size: 14px;
+			margin-bottom: 24px;
+		}
+	}
+
+	.spinner {
+		display: inline-block;
+		vertical-align: -25%;
+	}
+
+	@include breakpoint( "<660px" ) {
+		max-width: 100%;
+		border-radius: 0;
 	}
 }
 
@@ -21,10 +58,6 @@
 
 .is-group-editor .wp-content {
 	padding: 0;
-}
-
-.is-group-editor .email-verification-notice {
-	display: none;
 }
 
 .post-editor__inner {
